+++ src/argilla/server/schemas/v1/records.py
@@ -54,8 +54,12 @@ class RecordUpdate(UpdateSchema):
 
     @validator("metadata_", pre=True)
     @classmethod
-    def skip_nan_values(cls, metadata: Optional[Dict[str, Any]]) -> Optional[Dict[str, Any]]:
+    def prevent_nan_values(cls, metadata: Optional[Dict[str, Any]]) -> Optional[Dict[str, Any]]:
         if metadata is None:
             return metadata
+
+        for k, v in metadata.items():
+            if v != v:
+                raise ValueError(f"NaN is not allowed as metadata value, found NaN for key {k!r}")
 
         return {k: v for k, v in metadata.items() if v == v}  # By definition, NaN != NaN